+++ ndonnx/_core/_numericimpl.py
@@ -16,10 +16,12 @@ import numpy as np
 import ndonnx as ndx
 import ndonnx._data_types as dtypes
 import ndonnx._opset_extensions as opx
+import ndonnx.additional as nda
 from ndonnx._utility import promote
 
+from ._coreimpl import CoreOperationsImpl
+from ._interface import OperationsBlock
 from ._nullableimpl import NullableOperationsImpl
-from ._shapeimpl import UniformShapeOperations
 from ._utils import (
     binary_op,
     from_corearray,
@@ -35,7 +37,7 @@ if TYPE_CHECKING:
     from ndonnx._corearray import _CoreArray
 
 
-class NumericOperationsImpl(UniformShapeOperations):
+class _NumericOperationsImpl(OperationsBlock):
     # elementwise.py
 
     @validate_core
@@ -198,7 +200,7 @@ class NumericOperationsImpl(UniformShapeOperations):
     def isinf(self, x):
         if isinstance(x.dtype, (dtypes.Floating, dtypes.NullableFloating)):
             return unary_op(x, opx.isinf)
-        return ndx.full(x.shape, fill_value=False)
+        return ndx.full(nda.shape(x), fill_value=False)
 
     @validate_core
     def isnan(self, x):
@@ -400,9 +402,7 @@ class NumericOperationsImpl(UniformShapeOperations):
             return (ndx.arange(0, x != 0, dtype=dtypes.int64),)
 
         ret_full_flattened = ndx.reshape(
-            from_corearray(
-                opx.ndindex(ndx.asarray(x.shape, dtype=dtypes.int64)._core())
-            )[x != 0],
+            from_corearray(opx.ndindex(nda.shape(x)._core()))[x != 0],
             [-1],
         )
 
@@ -413,7 +413,7 @@ class NumericOperationsImpl(UniformShapeOperations):
                         ret_full_flattened._core(),
                         ndx.arange(
                             i,
-                            ret_full_flattened.shape[0],
+                            nda.shape(ret_full_flattened)[0],
                             x.ndim,
                             dtype=dtypes.int64,
                         )._core(),
@@ -454,20 +454,21 @@ class NumericOperationsImpl(UniformShapeOperations):
             from_corearray, opx.get_indices(x1._core(), x2._core(), positions._core())
         )
 
-        how_many = ndx.zeros(ndx.asarray(combined.shape) + 1, dtype=dtypes.int64)
+        combined_shape = nda.shape(combined)
+        how_many = ndx.zeros(combined_shape + 1, dtype=dtypes.int64)
         how_many[
-            ndx.where(indices_x1 + 1 <= combined.shape[0], indices_x1 + 1, indices_x1)
+            ndx.where(indices_x1 + 1 <= combined_shape[0], indices_x1 + 1, indices_x1)
         ] = counts
-        how_many = ndx.cumulative_sum(how_many, include_initial=True)
+        how_many = ndx.cumulative_sum(how_many, include_initial=False, axis=None)
 
-        ret = ndx.zeros(x2.shape, dtype=dtypes.int64)
+        ret = ndx.zeros(nda.shape(x2), dtype=dtypes.int64)
 
         if side == "left":
             ret = how_many[indices_x2]
-            ret[nan_mask] = ndx.asarray(x1.shape, dtype=dtypes.int64) - 1
+            ret[nan_mask] = nda.shape(x1) - 1
         else:
             ret = how_many[indices_x2 + 1]
-            ret[nan_mask] = ndx.asarray(x1.shape, dtype=dtypes.int64)
+            ret[nan_mask] = nda.shape(x1)
 
         return ret
 
@@ -494,7 +495,7 @@ class NumericOperationsImpl(UniformShapeOperations):
         # FIXME: I think we can simply use arange/ones+cumsum or something for the indices
         # maybe: indices = opx.cumsum(ones_like(flattened, dtype=dtypes.i64), axis=ndx.asarray(0))
         indices = opx.squeeze(
-            opx.ndindex(ndx.asarray(flattened.shape, dtype=dtypes.int64)._core()),
+            opx.ndindex(nda.shape(flattened)._core()),
             opx.const([1], dtype=dtypes.int64),
         )
 
@@ -502,7 +503,7 @@ class NumericOperationsImpl(UniformShapeOperations):
 
         values = from_corearray(ret_opd[0])
         indices = from_corearray(indices[ret_opd[1]])
-        inverse_indices = ndx.reshape(from_corearray(ret_opd[2]), x.shape)
+        inverse_indices = ndx.reshape(from_corearray(ret_opd[2]), nda.shape(x))
         counts = from_corearray(ret_opd[3])
 
         return ret(
@@ -535,7 +536,7 @@ class NumericOperationsImpl(UniformShapeOperations):
         if axis < 0:
             axis += x.ndim
 
-        _len = ndx.asarray(x.shape[axis : axis + 1], dtype=dtypes.int64)._core()
+        _len = ndx.asarray(nda.shape(x)[axis : axis + 1], dtype=dtypes.int64)._core()
         return _via_i64_f64(
             lambda x: opx.top_k(x, _len, largest=descending, axis=axis)[1], [x]
         )
@@ -544,7 +545,7 @@ class NumericOperationsImpl(UniformShapeOperations):
     def sort(self, x, *, axis=-1, descending=False, stable=True):
         if axis < 0:
             axis += x.ndim
-        _len = ndx.asarray(x.shape[axis : axis + 1], dtype=dtypes.int64)._core()
+        _len = ndx.asarray(nda.shape(x)[axis : axis + 1], dtype=dtypes.int64)._core()
         return _via_i64_f64(
             lambda x: opx.top_k(x, _len, largest=descending, axis=axis)[0], [x]
         )
@@ -566,13 +567,47 @@ class NumericOperationsImpl(UniformShapeOperations):
                 axis = 0
             else:
                 raise ValueError("axis must be specified for multi-dimensional arrays")
+
+        if isinstance(x.dtype, (dtypes.Unsigned, dtypes.NullableUnsigned)):
+            if ndx.iinfo(x.dtype).bits < 64:
+                out = x.astype(dtypes.int64)
+            else:
+                return NotImplemented
+        elif dtype == dtypes.uint64 or dtype == dtypes.nuint64:
+            raise ndx.UnsupportedOperationError(
+                f"Unsupported dtype parameter for cumulative_sum {dtype} due to missing kernel support"
+            )
+        else:
+            out = x.astype(_determine_reduce_op_dtype(x, None, dtypes.uint64))
+
         out = from_corearray(
             opx.cumsum(
-                x._core(), axis=opx.const(axis), exclusive=int(not include_initial)
+                out._core(),
+                axis=opx.const(axis),
+                exclusive=0,
             )
         )
-        if dtype is not None:
+
+        if dtype is None:
+            if isinstance(x.dtype, dtypes.Unsigned):
+                out = out.astype(ndx.uint64)
+            elif isinstance(x.dtype, dtypes.NullableUnsigned):
+                out = out.astype(ndx.nuint64)
+        else:
             out = out.astype(dtype)
+
+        # Exclude axis and create zeros of that shape
+        if include_initial:
+            out_shape = nda.shape(out)
+            out_shape[axis] = 1
+            out = ndx.concat(
+                [
+                    ndx.zeros(out_shape, dtype=out.dtype),
+                    out,
+                ],
+                axis=axis,
+            )
+
         return out
 
     @validate_core
@@ -707,7 +742,7 @@ class NumericOperationsImpl(UniformShapeOperations):
             and isinstance(x.dtype, dtypes.Numerical)
         ):
             x, min, max = promote(x, min, max)
-            if isinstance(x.dtype, dtypes._NullableCore):
+            if isinstance(x.dtype, dtypes.NullableCore):
                 out_null = x.null
                 x_values = x.values._core()
                 clipped = from_corearray(opx.clip(x_values, min._core(), max._core()))
@@ -806,17 +841,6 @@ class NumericOperationsImpl(UniformShapeOperations):
         )
 
     @validate_core
-    def make_nullable(self, x, null):
-        if null.dtype != dtypes.bool:
-            raise TypeError("'null' must be a boolean array")
-
-        return ndx.Array._from_fields(
-            dtypes.into_nullable(x.dtype),
-            values=x.copy(),
-            null=ndx.reshape(null, x.shape),
-        )
-
-    @validate_core
     def can_cast(self, from_, to) -> bool:
         if isinstance(from_, dtypes.CoreType) and isinstance(to, ndx.CoreType):
             return np.can_cast(from_.to_numpy_dtype(), to.to_numpy_dtype())
@@ -824,7 +848,7 @@ class NumericOperationsImpl(UniformShapeOperations):
 
     @validate_core
     def all(self, x, *, axis=None, keepdims: bool = False):
-        if isinstance(x.dtype, dtypes._NullableCore):
+        if isinstance(x.dtype, dtypes.NullableCore):
             x = ndx.where(x.null, True, x.values)
         if functools.reduce(operator.mul, x._static_shape, 1) == 0:
             return ndx.asarray(True, dtype=ndx.bool)
@@ -834,7 +858,7 @@ class NumericOperationsImpl(UniformShapeOperations):
 
     @validate_core
     def any(self, x, *, axis=None, keepdims: bool = False):
-        if isinstance(x.dtype, dtypes._NullableCore):
+        if isinstance(x.dtype, dtypes.NullableCore):
             x = ndx.where(x.null, False, x.values)
         if functools.reduce(operator.mul, x._static_shape, 1) == 0:
             return ndx.asarray(False, dtype=ndx.bool)
@@ -866,7 +890,7 @@ class NumericOperationsImpl(UniformShapeOperations):
 
     @validate_core
     def tril(self, x, k=0) -> ndx.Array:
-        if isinstance(x.dtype, dtypes._NullableCore):
+        if isinstance(x.dtype, dtypes.NullableCore):
             # NumPy appears to just ignore the mask so we do the same
             x = x.values
         return x._transmute(
@@ -877,7 +901,7 @@ class NumericOperationsImpl(UniformShapeOperations):
 
     @validate_core
     def triu(self, x, k=0) -> ndx.Array:
-        if isinstance(x.dtype, dtypes._NullableCore):
+        if isinstance(x.dtype, dtypes.NullableCore):
             # NumPy appears to just ignore the mask so we do the same
             x = x.values
         return x._transmute(
@@ -948,9 +972,10 @@ class NumericOperationsImpl(UniformShapeOperations):
         return ndx.full_like(x, 0, dtype=dtype)
 
 
-class NullableNumericOperationsImpl(NumericOperationsImpl, NullableOperationsImpl):
-    def make_nullable(self, x, null):
-        return NotImplemented
+class NumericOperationsImpl(CoreOperationsImpl, _NumericOperationsImpl): ...
+
+
+class NullableNumericOperationsImpl(NullableOperationsImpl, _NumericOperationsImpl): ...
 
 
 def _via_i64_f64(